+++ src/main/java/org/springframework/aop/AdvisedSupport.java
@@ -9,7 +9,7 @@ import org.aopalliance.intercept.MethodInterceptor;
 public class AdvisedSupport {
 
 	//是否使用cglib代理
-	private boolean proxyTargetClass = false;
+	private boolean proxyTargetClass = true;
 
 	private TargetSource targetSource;