+++ src/test/java/org/springframework/test/bean/A.java
@@ -8,6 +8,8 @@ public class A {
 
 	private B b;
 
+	public void func(){}
+
 	public B getB() {
 		return b;
 	}