+++ src/test/java/org/springframework/test/ioc/CircularReferenceWithoutProxyBeanTest.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Java6Assertions.assertThat;
 public class CircularReferenceWithoutProxyBeanTest {
 
 	@Test
-	public void testCircularRefreence() throws Exception {
+	public void testCircularReference() throws Exception {
 		ClassPathXmlApplicationContext applicationContext = new ClassPathXmlApplicationContext("classpath:circular-reference-without-proxy-bean.xml");
 		A a = applicationContext.getBean("a", A.class);
 		B b = applicationContext.getBean("b", B.class);